move conserved quantity computation to observables
