+++ main/src/propagator/ve_hydro.hpp
@@ -192,13 +192,10 @@ public:
         timer.step("Timestep");
         computePositions(first, last, d, domain.box());
         timer.step("UpdateQuantities");
-        computeTotalEnergy(first, last, d);
-        timer.step("EnergyConservation");
         updateSmoothingLength(first, last, d, ng0_);
         timer.step("UpdateSmoothingLength");
 
         timer.stop();
-        this->printIterationTimings(domain, d);
     }
 
     //! @brief configure the dataset for output by calling EOS again to recover rho and p
